requirements-dev.txt
console_log==0.2.10
flake8-coding==1.3.0
flake8-commas==2.0.0
flake8-future-import==0.4.5
flake8-import-order==0.18
flake8-quotes==1.0.0
flake8==3.5.0
flask-cors==3.0.3
ipdb==0.11
mysqlclient==1.3.13
psycopg2-binary==2.7.5
pycodestyle==2.4.0
pylint==1.9.2
redis==2.10.6
statsd==3.2.2
tox==3.1.2
